pyproject.toml
[project]
name = "swotbot"
version = "2025.06.21.0001"
description = "SWOTbot is an AI-powered tool that generates comprehensive SWOT (Strengths, Weaknesses, Opportunities, Threats) analyses for your project outlines. It uses AI agents to analyze your project description and provide professional, factual, and straightforward insights."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.6.3",
    "duckduckgo-search>=8.0.4",
    "fastapi>=0.115.13",
    "hatchling>=1.27.0",
    "mistralai>=1.8.2",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "tavily-python>=0.7.7",
    "typer>=0.16.0",
    "uvicorn>=0.34.3",
]

[project.scripts]
swotbot = "swotbot:main.run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
